Some cleanup and testing the MCAPBufferedFileChannelInput
