+++ scs2-session-logger/build.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
    id("us.ihmc.ihmc-build")
-   id("us.ihmc.ihmc-ci") version "8.3"
 }
 
 ihmc {
@@ -22,4 +21,5 @@ mainDependencies {
 }
 
 testDependencies {
+   api("org.junit.jupiter:junit-jupiter-api:5.10.1")
 }
